backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
channels>=4.0.0
channels-redis>=4.1.0
psycopg[binary]>=3.1.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
django-cors-headers>=4.3.1
